+++ classpath.js
@@ -1,0 +1,6 @@
+/**
+ * Copyright 2017 aixigo AG
+ * Released under the MIT license.
+ * https://opensource.org/licenses/MIT
+ */
+module.exports = [ `${__dirname}/java` ];