+++ java/com/aixigo/pdfreactor_loader/LoggingHandler.java
@@ -1,3 +1,8 @@
+/**
+ * Copyright 2017 aixigo AG
+ * Released under the MIT license.
+ * https://opensource.org/licenses/MIT
+ */
 package com.aixigo.pdfreactor_loader;
 
 import java.util.logging.Handler;