+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#2](https://github.com/aixigo/pdfreactor-loader/issues/2): Improved adding classpath entries and added workaround notice
 - [#1](https://github.com/aixigo/pdfreactor-loader/issues/1): PDFreactor log messages are now written to console